pyproject.toml
# ===================================================================
# VDM - Project Configuration File
# ===================================================================
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "vdm-server"
version = "0.0.1"
description = "A multiplayer, AI-driven storytelling game server."
requires-python = ">=3.11"

dependencies = [
    # Server & API Framework
    "fastapi",
    "orjson",
    "pydantic-settings",
    "uvicorn",

    # Networking
    "httpx",

    # User Management
    "passlib[bcrypt]",

    # AI & Machine Learning
    "chonkie[st]",
    "chromadb",
    "hf_xet",
    "kokoro",
    "misaki",
    "sentence-transformers",
    "soundfile",
    "torch",
    "torchaudio",

    # Utilities
    "PyYAML",
    "rich",
]

[project.optional-dependencies]
gpu = [
    "torch",
    "torchaudio",
]

[tool.setuptools.packages.find]
where = ["src"]